Adding support for calling garbage collection during simplification.
That the memory usage usually peaks during simplification has been a
weakness for quite some time. It is still the case even with this change,
but matters are improved considerably by making sure that the set of
removed clauses does not grow arbitrarily. The change comes at a price of
somewhat reduced performance in cases where many garbage collections are
triggered and it makes sense to monitor this behavior in the future.