Make the default binary statically linked.
The default binary was dynamically linked (*dynamic* target instead of
*release* target), but this meant that the default binary failed to
run if LD_LIBRARY_PATH was not set, when PREFIX was set to a
non-default value. In other words, installing minisat into your home
dir resulted in a binary that would not run out of the box.  The
problem is that the dynamically linked minisat binary depends on
libminisat.

I ran into this while trying to install tip from source: it needed a
newer version of minisat than my package manager provided.