output: minimal change for SAT competition
While Minisat prints the output state without the 's ' prefix, and prints
models without the 'v ' prefix, the SAT competition, and related events
require those prefixes. Furthermore, Minisat does not print the model to
the screen by default. Finally, any other data that is printed by Minisat
is not prefixed by a 'c ', so that output scanners might be confused.

To make the output of the solver ready for competitions, add the 's ' and
'v ' prefixes, and 'c ' prefix to the messages that are printed
independently of the set verbosity.