output: switch printing model
The original variant of Minisat prints the model only to the output file,
but did not print it to screen. For competitions, the model has to be
printed as well, hence, it was added to the previous commit.

To be able to restore the old behavior, add a CLI parameter "-model", that
controls whether the values of the model are printed. When specifying
'-no-model' as a parameter, no model is printed.