proof: add essential requirements
This commit adds printing proofs to a file. To allow printing, the API of
the solver is extended by three functions that open a proof file, extend
the proof by adding or deleting a clause, or finalize the proof and close
the file again. For the simplifying solver, an operation mode flag is
introduced, which indicates that the solver currently parses the input.

During operation, the solver only uses the extendProof method. This allows
to easily adapt the format in the future.

This commit is based on the patch by Marijn Heule that modified an older
version of Minisat (v2.2.0) to comply with the SAT competition rules:
http://www.satcompetition.org/2013/files/MinisatHack.patch
In this commit, only the proof generation is implemented. And furthermore
a bug in the addClause method is fixed, recognizing tautologies correctly.