IntSet: allow conversion to Vec
To support programs that use Minisat as a backend, allow to convert an
IntSet into the underlying Vec container. This then enables to use the
conflict member of the Solver class either as IntSet, or as Vec, so that
both the recent definition, as well as the older definition of this
variable are supported.