proof: fix variable elimination
While for data structure maintenance it is benefical that clauses are
removed from lists before these lists are used to execute further
computation, proofs require to add the new clauses before the old ones
can be deleted.

Therefore, keep the computation as is, and only consider the clauses to
be removed in another removal round.