SWDiA5BY: add solver
This solver has been constructed based on the partial steps that are
available on the web page of cominisatps. While the description
mentions several patches, only 2 of them have been available.

Note, despite the web page claims that the patches turn Minisat 2.2.0
into Glucose 2.3, the namespace change did not make it, so that this
commit also undoes this change back to the namespace Minisat.