almostPure: add feature
This modification allows to modify the picked polarity, in case all
non-unit non-satisfied clauses contain a literal of the same polarity.
The check is performed before propagating unit clauses.

While this analysis is incomplete, it actually covers formulas that
can be satisfied by a pure strategy if there would not be units.