LCM: add reverse feature
In case vivification works for the first order of the clause, try the
reverse as well. In case the initial clause cannot be simplified via
vivification, there is no need to try. However, in case some
simplification was doable, trying the reversed clause is a natural
choice, given that any other order would potentially work as well.