dec: make VSIDS switch deterministic
Currently, we switch from the selected heuristic back to VSIDS after
about 2500 seconds. However, this results in dependencies on the used
architecture, and non-deterministic behaviour that cannot e reproduced
easily.

To achieve reproducability of results again, we drop the timer to
trigger the switch, and set values for propagations and conflicts so
that the switch is performed.

This furthermore gets rid of the global variable.

As we are at it, fix the way the switch kicks in when no trigger is
given.