ipasir: add initial interface
This file allows make MiniSat available as IPASIR backend, so that
the solver can be used as SAT backend.

Note: in addition to the general ipasir interface, we also provide
the ipasir_solve_final method, which can be used in case the user
already knows that the current calls to the SAT solver is the final
call, which allows full formula simplification.

The file has been provided by Armin Biere to be used with the MIT
license.