asan: fix out of bound access
When making a variable a decision variable, we need to have the required
space in all the other structures first.

This bug might result in variables not ending up in the heap, which might
lead to models that do not have that variable, as it might not end up in
the actual heap.