format: add first version
This format file can be used to format the source code of the solver to a
common style. Most importantly, this allows to merge changes to other
variants of MiniSat more easily, by following the following steps:

1. Take the base solver, apply the format
2. Take the modification, apply the format
3. Create the diff between the formatted sources
4. Format the target solver with the same format
4. Turn the diff into a patch, and port it the target solver

The current format has been tested, and is close to what MiniSat is using
at the moment. The above procedure would work with any other format as well.