namespace: make compile time config
To be able to compile MergeSAT as a SAT backend for different tools,
allow to be flexible on the used namespace. This allows to use the
solver as a MiniSat or Glucose replacement, as well as a parallel
SAT backend when MiniSat or Glucose are already used in a backend.

To not use the namespace Minisat, define a value for the variable
MERGESAT_NSPACE during compile time. To only process the default case
once, the commonly used header file XAlloc defines the default value,
and had to be included only one more time.