decision: toggle more stable
When toggling decision heuristics without updating decision heaps is okay
as long as the solver is not used in an incremental mode.

In incremental mode however, we might solve a formula, and call the solver
again afterwards. This results in switched back to VSIDS, eventually
without rebuilding the heap fully. This can lead to reporting a satisfiable
answer for unsatisfiable formulas, as has happened when using a buggy
implementation as a SAT backend for the OpenWBO SAT solver.

This commit introduced a method that allows to toggle the decision heaps
without running into this problem.