docker: allow to compile and run
This commit extends the README with instructions to use docke for
compiling the solver, as well as running the solver if required.

Note, the provided docker infrastructure furthermore links the
solver against a modified glibc, which enables using transparent
huge pages via madvise. In average, this can improve the runtime of
the solver by 10%, depending on the input up to 20% or none at all.
For long running formulas with a high memory consumption, the expected
speedup is around 10%.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>