incremental: do not clean lookup tables
When switching decision heuristics, the current code cleans lookup tables,
but in a future iteration of incremental solving might add more variables.
After a clear, using push to add more capacity is the wrong combination.
Furthermore, not consuming the data for variables that have been present
already is the wrong move. Hence, we do not clean the structures.

Compared to the previous implementation, the memory of the variables is
preserved anyways, as the usual "clear()" does not actually free the
memory.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>