decision: fix toggling heaps
Until now, we always moved heaps to VSIDS, independently of which heap
is actually used for picking decision variables. In the worst case, this
results in the solver reporting SAT for unsatisfiable formulas, namely
in combination with partial restarts, where the decision level is not
zero and there are pending assigned variables.

Fixes: 7404f17 (decision: toggle more stable)

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>