search: check for variables to select
The variables that should be available as decision variables should
cover all unassigned variables. To be able to quickly ignore variables
that are no decision variables, add a new variables that stores the
maximal size when decision heaps are build, as this size represents the
number of variables that has to be assigned to prove a formula satisfiable.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>