restart: allow random partial
To test partial restarts, jumping to a random level helps to not jump to
the top level but to jump back partially. This behavior helps for
testing the combination of partial restarts with other techniques. As
the randomization is not fully random, but based on a random seed, a
failure can be reproduced.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>