fuzzing: add script to fuzz several configurations
When fuzzing the default configuration, many states in the solver might
not be covered due to the involved scheduling. To work around these
limitations, also fuzz configurations in debug mode, as well as with a
reduced time-span in the schedule. This allows to cover more states
faster. Furthermore, use check-sat during fuzzing, to make sure we can
identify issues in the solver.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>