libz: remove gzip support
This change drops the support for reading from compressed files directly.
The feature will be revived with a reading from pipe approach, which allows
to easily extend the list of supported compression.

To be able to support solver consumers that rely on the presence of libz
for parsing, allow to reactivaze libz via the compile time flag USE_LIBZ.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>