UP,enqueue: ask for level parameter
When using the enqueue function, the default decision level to put
literals in is 0. To make sure we actually want to set this level, make
the level parameter non-default, i.e. enforce setting the required level
explicitly.

While running through the code, we spotted places where 0 was used
wrongly. Now, the solver use the enqueue function as expected.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>