ci: add mini competition
We want to get a basic understanding of the performance of the solver
when applying changes. Hence, use a few instances of the 2020
competition to run a small benchmark test.

This script currently does not test the output of the solver for
correctness, i.e. models and proofs are not checked at the moment.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>