inprocess: respect different marks
More recent solvers that depend on MiniSat and Glucose do not use mark()
for labeling deletion, but also label clause tiers. Hence, we cannot
easily port mark() modifiers, but instead have to make sure we compare
against 1 to fetch deletion information.

This change catches this behavior, and uses the correct label in the
inprocessing routine.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>