backtracking: avoid loops
In edge cases, a conflict clause is reported for a decision level, which
actually would be asserting on a higher decision level already. After
partially backtracking, and further propagation, we can hit the very
same conflict again. These steps can be repeated in a loop.

This change modifies the backjumping, by making sure to not hit the same
conflict multiple times. In case we hit the same conflict again, with the
same property of having a single literal on its highest decision level,
we will jump back to properly hit this conflict during propagation
afterwards. This way, loops can be avoided.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>