ci,configurations: split release and debug
For debug builds, test execution is costly. Hence, we should reduce the
number of formulas to be tested per configuration.

This change uses the number of formulas to be tested per tool
configuration. While being at it, we also fix the style of the script
using:

shfmt -i 4 -w <script>

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>