search,heap: rebuild only if necessary
In case we call the rebuildOrderHeap function, but the heap currently
holds more variables than are unassigned in the formula, and we are on
the root decision level, do not rebuild the heap at all. In this
situation, the heap should contain all variables that are still
unassigned.

In debug mode, this property can be checked, before we use the short cut.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>