signal: exit interrupt on 2nd try
Some parts of the solver implementation take a while until they call
the check for interrupts again. This change allows to terminate the
program in interactive mode either way.

In case we are interrupted due to a CPU limit, exit on the first try
with the same exit code as the timeout command.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>