ci: timeout ipasir calls, add debug
There are calls during incremental solving testing that take very long.
We add a timeout check to these tests, so that we also detect timeouts,
which hint at performance issues for incremental solving.

Integrating the timeout checks allows to identifies these regressions
early during development. The default timeout is set to 3600 seconds.

Furthermore, to catch more errors, add another iteration that uses
the debug variant of MergeSat as backend, so that assertions of the
solver are active.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>