mini-competition: add dryrun modes
Instead of actually running the solvers, only print the commands
that would be executed. To reduce noise on systems that use network
file systems, access the benchmark before actually passing it to
the solver.

The output files for the solver are generated based on the solver,
it's command line arguments, as well as the path of the benchmark.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>