trail-saving,fix: keep conflicts structure sane
When we pass a conflict to code analysis, we need to make sure that
the first literal is the conflicting literal, as this literal is
excluded in the analysis method during resolution. Using a wrong
literal can easily result in wrong learned clauses, and wrong
outcomes of the search.

To make sure we detect issues early, add further assertions to
make sure we are not violating the internal state.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>