hordesat: implement backend hooks
For integration into hordesat, we need to provide functionality to
add clauses to the solver, as well as to share clauses with other
solver incarnations. Hence, add the relevant functionality to the
solver.

Additionally to the changes to MiniSat in hordesat, also allow to limit
the clauses that will be shared based on size. Furthermore, allow to
control whether solving instances are sharing or receiving clauses at
all. This allows to keep known configurations stable, i.e. the first
configuration in a parallel set could always represent the best
sequential configuration. By not receiving clauses, the parallel solver
will be at least as powerful as this configuration (modulo the over-
head of sharing compute resources).

Furthermore, instead of implementing the configuration diversification
in the calling instance, add a stub of a function to diversify the
solver configuration inside the solver. This way, we make sure we can
easily grow diversification without changing multiple tools.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>