sls,rephase,incremental: disable for assumptions
Apparently, the current implementation gets SLS for MergeSat wrong.
An obvious fix is to disable SLS for incremental solving for now.

As rephasing is based on the information of SLS, also disable
rephasing. As there is no single controlling variable for rephasing,
we introduce one, as well as a CLI option, as well.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>